+++ Core.lua
@@ -35,25 +35,6 @@ function RaidFrameSettings:OnInitialize()
     self:RegisterChatCommand("rfs", "SlashCommand")
     self:RegisterChatCommand("raidframesettings", "SlashCommand")
     self:RegisterEvent("PLAYER_LOGIN", "LoadGroupBasedProfile")
-
-    -- minimap icon
-    local minimapIconConf = LDB:NewDataObject("RaidFrameSettings_Excorp_Fork", {
-        type = "launcher",
-        icon = "Interface\\AddOns\\RaidFrameSettings_Excorp_Fork\\Textures\\Icon\\Icon.tga",
-        OnClick = function(self, button)
-            if button == "LeftButton" then
-                RaidFrameSettings:SlashCommand()
-            elseif button == "RightButton" then
-                RaidFrameSettings:ShowMinimapIcon(false)
-            end
-        end,
-        OnTooltipShow = function(tooltip)
-            tooltip:SetText(L["RaidFrameSettings"])
-            tooltip:AddLine(L["left button - Toggles the options window."], 1, 1, 1)
-            tooltip:AddLine(L["right button - Hides the minimap icon."], 1, 1, 1)
-        end,
-    })
-    LDBIcon:Register("RaidFrameSettings_Excorp_Fork", minimapIconConf, self.db.profile.MinorModules.minimapIcon)
 end
 
 function RaidFrameSettings:SlashCommand()
@@ -81,7 +62,6 @@ function RaidFrameSettings:OnEnable()
             module:Enable()
         end
     end
-    self:ShowMinimapIcon()
     self:RegisterEvent("GROUP_ROSTER_UPDATE", "CheckGroupType") --GroupType.lua
 end
 
@@ -98,7 +78,6 @@ end
 function RaidFrameSettings:UpdateModule(module_name)
     self:DisableModule(module_name)
     self:EnableModule(module_name)
-    self:ShowMinimapIcon()
 end
 
 function RaidFrameSettings:ReloadConfig()
@@ -106,17 +85,6 @@ function RaidFrameSettings:ReloadConfig()
     self:GetProfiles()
     self:LoadUserInputEntrys()
     self:Enable()
-end
-
-function RaidFrameSettings:ShowMinimapIcon(show)
-    if show ~= nil then
-        RaidFrameSettings.db.profile.MinorModules.minimapIcon.hide = not show
-    end
-    if RaidFrameSettings.db.profile.MinorModules.minimapIcon.hide then
-        LDBIcon:Hide("RaidFrameSettings_Excorp_Fork")
-    else
-        LDBIcon:Show("RaidFrameSettings_Excorp_Fork")
-    end
 end
 
 --Addon compartment