+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # **Changelog**
+### Version [2.25.0] - 2024-03-29
+#### Changed
+* Reworked Buffs for classic to match feature parity with retail.
+
+#### Fixes
+* Fixed an issue with the roster cache building that affected the classic version if the user was not playing with KeepGroupsTogether, which also fixed an issue with the cache building when changing the edit mode layout between Separate Groups and Combine Groups.
+
+### Version [2.24.2] - 2024-03-29
+#### Added
+* Minimap icon
+
+#### Changed
+* Once again, the Options Window will re-appear after combat ends. I don't know why I removed it in the first place.
+
 ### Version [2.24.1] - 2024-03-29
 #### Added
 * FrameColor Skin